+++ tomcat/src/main/java/org/apache/coyote/http11/request/header/Cookies.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.header;
 
 import java.util.Arrays;
 import java.util.HashMap;
