+++ tomcat/src/main/java/org/apache/coyote/http11/request/header/Location.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.header;
 
 import java.net.URL;
 import java.nio.file.Path;
