+++ tomcat/src/main/java/org/apache/coyote/http11/request/header/Method.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.header;
 
 public enum Method {
     GET,
