+++ tomcat/src/main/java/org/apache/coyote/http11/request/header/UserSession.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.header;
 
 import nextstep.jwp.model.User;
