+++ tomcat/src/test/java/org/apache/coyote/http11/response/CookiesTest.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.response;
 
-import org.apache.coyote.http11.request.Cookies;
+import org.apache.coyote.http11.request.header.Cookies;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
