+++ tomcat/src/test/java/org/apache/coyote/http11/response/LocationTest.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.response;
 
-import org.apache.coyote.http11.request.Location;
+import org.apache.coyote.http11.request.header.Location;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
