+++ tomcat/src/test/java/org/apache/coyote/http11/response/MethodTest.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.response;
 
-import org.apache.coyote.http11.request.Method;
+import org.apache.coyote.http11.request.header.Method;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
