+++ tomcat/src/test/java/org/apache/coyote/http11/response/ParametersTest.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.response;
 
-import org.apache.coyote.http11.request.Parameters;
+import org.apache.coyote.http11.request.body.Parameters;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
