+++ tomcat/src/test/java/org/apache/coyote/http11/response/UserSessionTest.java
@@ -1,7 +1,7 @@
 package org.apache.coyote.http11.response;
 
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.request.UserSession;
+import org.apache.coyote.http11.request.header.UserSession;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
